move `lib.rs` and `main.rs` back
