--- src/explore/main.rs
@@ -1,6 +1,0 @@
-use nu_plugin::{serve_plugin, MsgPackSerializer};
-use nu_plugin_explore::Explore;
-
-fn main() {
-    serve_plugin(&mut Explore {}, MsgPackSerializer {})
-}
